Delete `application.hbs`
When you are done experimenting, go ahead and delete the file
`app/templates/application.hbs`. We won't be needing this for a while,
so we will start afresh and add it back later when there we have a need
for it.

If you still have your browser tab open, as you delete the file, your
tab should automatically reload into a blank page, reflecting the fact
that we no longer have an application template in our app.