Add `teaching-tomster.png`
To match mockup from our designer, we will also need to download the
tomster image, which was referenced from our CSS file:

```css
.tomster {
  background: url(../assets/images/teaching-tomster.png);
  /* ...snip... */
}
```

As we learned earlier, the Ember convention is to place your source code
in the `app` folder. For other assets like images and fonts, it is
conventional to put them in the `public` folder. So we will download the
image into `public/assets/images/teaching-tomster.png`.

Ember CLI and the development server understands these folder
conventions and will automatically make them available to the browser.
You can confirm this by navigating to
`http://localhost:4200/assets/images/teaching-tomster.png`. It should
also show up in the welcome page we have been working on, but a manual
refresh may be needed for the browser to pick up the new file.