Add `about` route
With our first page down, let's add another one!

This time, we would like the page to be served on the `/about` URL. To
do this, we will need to tell Ember our plan to add a page at that
location, otherwise Ember will think we have visited an invalid URL!

The place to manage what pages are available is the _router_. Go ahead
and open `app/router.js` and make the following change:

<!-- TODO: show diff -->

This adds a _route_ named "about", which is served at the `/about` URL
by default. With that in place, we can create a our new page under
`app/templates/about.hbs` with the following content:

<!-- TODO: show diff -->

To see this in action, navigate to `http://localhost:4200/about`. With
that, our second page is done!