Add `contact` route
We are on a roll!

While we are at it, let's add our third page. This time, things are a
little bit different. Everyone at the company calls this the "contact"
page, but old website we are replacing already has a similar page, but
served at the legacy URL `/getting-in-touch`.

We want to keep the existing URLs for the new website, but
`getting-in-touch` is a mouthful to type and to say-out-loud all the
time.  Fortunately, we can have the best of the both worlds:

<!-- TODO: show diff -->

We can accomplish this by explicitly specifying a path for the route.
This allows us to keep the legacy URL, but use the new, shorter name for
the route, and therefore the template filename as well.

Ember comes with strong _convensions_ and sensible defaults – after all,
if we are strating from scratch we wouldn't mind the default `/contact`
URL. However, if the defaults doesn't work for us, it is not a problem
at all to customize Ember for our needs!

Once you have make the changes above, we should have the new page
available to us at `http://localhost:4200/getting-in-touch`.

Congratulations, you have completed your training as a master
page-crafter!