Use `<LinkTo>` to connect the pages
Actually, there is one more thing. We put so much effort into making
these pages, we need to make sure people can find them! Of course, the
way we do that on the web is using hyperlinks!

Since Ember offers great support for URLs out-of-the-box, we _could_
just link our pages together using the `<a>` tag with the appropiate
`href`. However, clicking on those links would require the browser to
make a trip back to the server and load everything from stratch again,
known as a _full-page refresh_.

With Ember, we can do better than that! Instead of the plain-old `<a>`
tag, Ember provides an alternative called `<LinkTo>`. Here is how you
would use it:

<!-- TODO: show diff -->

Go ahead and try it in the browser! Yay, no more manually changing the
URL!

There is quite a bit going on here, so let's break it down.

`<LinkTo>` is an example of a _component_ in Ember – you can tell them
apart from regular HTML tags because they start with an uppercase
letter. Along with regular HTML tags, components are a key building
block for building up your app's user interface.

We have a lot more to say about components later, but for now, you can
think of them as a way to provide  _custom tags_ to suppliment the
built-in ones that came with the browser.

The `@route=...` part is how we pass _arguments_ into the component.
Here, we use this to specify _which_ route we want to link to. Note that
this should be the _name_ of the route, not the path, which is why we
specified "about" instead of "/about", and "contact" instead of
"/getting-in-touch".

Finally, components can take the usual HTML attributes as well. Here, we
added a "button" class for styling purposes, but we could also specify
other attributes, such as the ARIA `role` attribute, as we see fit.
These are passed without the `@` symbol (`class=...` as opposed to
`@class=...`), so that Ember will know they are just regular HTML
attributes.

Under-the-hood, `<LinkTo>` generates a regular `<a>` tag for us with the
appropiate `href` for the specific route. This allows for perfect interop for
screenreaders, and it also allows your users to bookmark the link or open it in
a new tab.

However, when clicking one of these special links, Ember will intercept
the click, renders the content for the new page and updates the URL, all
performed locally without having to wait for the server, thus avoiding a
full page refresh.

We will learn more about how all of this work soon. In the meantime, go ahead a
give that a try. Did you notice how snappy it was?