Refactor into `application.hbs`
Before we move on to the next feature, there is one more thing we could
clean up.

Since the `<NavBar>` is used for site-wide navigation, it really needs
to be displayed on _every_ page in the app. So far, we have been adding
the component on each page manually. This is is a bit error-prone, as we
can easily forget to do so when adding a new page next time.

We could solve this problem by moving the nav-bar into a special
template called `application.hbs`. You may remember that it was
generated for us when we first created the app. It's time to bring it
back!

This template is special in that it does not have its own URL and cannot
be navigated to on its own. Instead, it is used for specifying a common
layout that is shared by every page in your app. This is a great place
to put site-wide UI elements, like a nav-bar and a site footer.

While we are at it, we will also add a container element that wraps
around the whole page, as requested by our designer for styling
purposes.

```handlebars {data-filename="app/templates/application.hbs"}
<div class="container">
  <NavBar />
  <div class="body">
    {{outlet}}
  </div>
</div>
```

```handlebars {data-filename="app/components/index.hbs" data-diff="-1"}
<NavBar />
<Jumbo>
  <h2>Welcome to Super Rentals!</h2>
  <!-- ...snip... -->
</Jumbo>
```

```handlebars {data-filename="app/components/about.hbs" data-diff="-1"}
<NavBar />
<Jumbo>
  <h2>About Super Rentals</h2>
  <!-- ...snip... -->
</Jumbo>
```

```handlebars {data-filename="app/components/contact.hbs" data-diff="-1"}
<NavBar />
<Jumbo>
  <h2>Contact Us</h2>
  <!-- ...snip... -->
</Jumbo>
```

Much nicer. As confirmed by our test suite, everything still works after
our refactor. We are ready to move on to the next feature!