Add `<Rental>` component
It's time to finally work on the rentals listing!

Let's start by creating a `<Rental>` component. This time, we will use
the component generator to create the template and test file for us.

```
ember generate component rental
```

We will _hard-code_ the rental property's details for now, and replace
it with the real data from the server later on. We will write a test to
ensure all the details are present.

<!-- TODO: show diff -->

Finally, let's invoke this a couple of times from our index template to
populate the page.

<!-- TODO: show diff -->

Things are looking pretty convincing already, not bad for just a little
bit of work!