Add `<Rental::Image>` component
Next, let's add the image for the rental property. We will use the
component generator for this again.

```
ember generate component rental/image
```

This time, we had a slash in the component's name. This resulted in the
component being created at `app/components/rental/image.hbs` instead of
at the root of `app/components`. This component can be invoked as
`<Rental::Image>`.

Components like these are known as _namespaced_ components. They allow
us to organize our components by folders according to the purpose. This
is completely optional – namespaced components are not special in any
way.

Let's edit the component's template:

```handlebars {data-filename="app/components/rental/image.hbs"}
<div class="image">
  <img ...attributes>
</div>
```

Instead of hard-coding a specfic `src` and `alt` attributes on the
`<img>` tag, we opted for the `...attributes` keyword instead, sometimes
called the _"splattributes"_ syntax. This allows arbitrary HTML
attributes to pass in when invoking this component:

```handlebars {data-filename="app/components/rental.hbs" data-diff="+2,+3,+4,+5,+6"}
<article class="rental">
  <Rental::Image
    src="https://upload.wikimedia.org/wikipedia/commons/c/cb/Crane_estate_(5).jpg"
    alt="A picture of Grand Old Mansion"
  />

  <div class="details">
    <!-- ...snip... -->
  </div>
</article>
```

We specified a `src` and `alt` HTML attribute here, which will be passed
along to the component and attached to the element where `...attributes`
is specified. You can think of this as similar to `{{yield}}`, but for
HTML attributes instead of content. In fact, we have already used this
feature earlier when we passing a `class` attribute to `<LinkTo>`.

That way, our `<Rental::Image>` component is not coupled to any specific
rental property on the site. Of course, the problem still exist, we
simply moved it to the `<Rental>` component, but we will deal with that
soon. We will try to limit all the hard-coding to the `<Rental>`
component, so that we will have an easier time cleaning it up when we
switch to fetching real data.

In general, it is a good idea to add `...attributes` to the primary
element in your component. This will allow for maximum flexibility as
the invoker may need to pass along classes styling or ARIA attributes
to improve accessibility.

Let's write a test for our component!

<!-- TODO: show diff -->

Finally, we should also update the tests for `<Rental>` to confirm that
we successfully invoked `<Rental::Image>`.

<!-- TODO: show diff -->

Because we already tested `<Rental::Image>` extensively on its own, we
can omit the details here and keep the assertion to the bare minimum.
That way, we won't have to _also_ update the `<Rental>` tests whenever
we make any changes to `<Rental::Image>`.