Add `RentalImageComponent` class
So far, all the components we have written are purely _presentational_ -
they are simply reusable snippets of markup. That is cool, but in Ember,
components can do so much more!

Sometimes, you want to associate some _behavior_ with your components to
do more. For example, `<LinkTo>` can respond to clicks by changing the
URL and navigating us to a different page.

Here, we are going to do just that! We are going to implementing the
"View Larger" and "View Smaller" functionality. Specifically, we want to
give our users the ability to view a larger version of the property's
image by clicking on it, and when they are done, they can click on it
again to go back to the smaller version.

In other words, we want a way to _toggle_ the image between one of the
two _states_. To do that, we need a way for the component to store and
be aware which of the two possible states it is currently in.

Ember allows us to optionally associate JavaScript code with a component
for exactly this purpose. We can generate a JavaScript file for our
`<Rental::Image>` component by running the `component-class` generator:

```
ember generate component-class rental/image
```

This generated a JavaScript file with the same name as our component's
template at `app/components/rentals/image.js`. It contains a _JavaScipt
class_, subclassing from `@glimmer/component`.

> Zoey says: `@glimmer/component`, or _Glimmer component_, is one of the
> several component classes available. They are a great starting point
> whenever you want to add behavior to your components. In this
> tutorial, we will be using Glimmer components exclusively.
>
> In general, Glimmer components should be used whenever possible.
> However, you may also see `@ember/components`, or _classic
> components_, being used in older apps. You can tell them apart from
> looking at their import path, which is helpful for looking up the
> respective documentation, as they have different and incompatible
> APIs.

Ember will create an _instance_ of the class whenever our component is
invoked. We can use that to store our state:

```js {data-filename="app/components/rental/image.js" data-diff="+4,+5,+6"}
import Component from '@glimmer/component';

export default class RentalImageComponent extends Component {
  constructor(...args) {
    super(...args);
    this.isLarge = false;
  }
}
```

Here, in the component's _constructor_, we _initialized_ the _instance
variable_ `this.isLarge` to the value `false`, as this is the default
state that we want for our component.

Let's update our template to use this state we just added:

```handlebars {data-filename="app/components/rental/image.hbs"}
{{#if this.isLarge}}
  <div class="image large">
    <img ...attributes>
    <small>View Smaller</small>
  </div>
{{else}}
  <div class="image">
    <img ...attributes>
    <small>View Larger</small>
  </div>
{{/if}}
```

The `{{#if ...}}...{{else}}...{{/if}}` _conditional_ syntax allows us to
render different content. We also have access to the component's
instance variables from the template. Combining these two features, we
can render either the small or large version of the image accordingly.

We can verify this is working by temporarily changing the initial value
in our JavaScript file. If we change `app/components/rental/image.js` to
initialize `this.isLarge = true;` in the constructor, we should see the
large version of the property image in the browser. Cool!

Since this pattern of initializing instance variables from the
constructor is pretty common, there is a more concise syntax for it:

```js {data-filename="app/components/rental/image.js" data-diff="+4,-5,-6,-7,-8"}
import Component from '@glimmer/component';

export default class RentalImageComponent extends Component {
  isLarge = false;
  constructor() {
    super(...arguments);
    this.isLarge = false;
  }
}
```

This does exactly the same thing as before, but much shorter!