Refactor `<Rental::Image>` to use inline `{{if}}`
Let's clean up our template before moving on. We have introduced a lot
of duplication when we added the conditional in the template. If we look
closely, the only thing that is different between the two blocks are:

1. The presence of the "large" CSS class on the `<button>` tag.
2. The "View Larger"/"View Smaller" text.

These changes are buried deep within the large amount of duplicated
code. We can reduce the duplication by using `{{if}}`  _expression_
instead:

```hbs {data-filename="app/components/rental/image.hbs"}
<button type="button" class="image {{if this.isLarge "large"}}" {{on "click" this.toggleSize}}>
  <img ...attributes>
  {{#if this.isLarge}}
    <small>View Smaller</small>
  {{else}}
    <small>View Larger</small>
  {{/if}}
</button>
```

The expression version of `{{if}}` takes two arguments. The first
argument is the _condition_. The second argument is what the expression
should evaluates into if the condition is true.

Optionally, it can also take a third argument for what the expression
should evaluate into if the condition is false. This means we could
alternatively rewrite the button label like so:

```hbs {data-filename="app/components/rental/image.hbs"}
<button type="button" class="image {{if this.isLarge "large"}}" {{on "click" this.toggleSize}}>
  <img ...attributes>
  <small>View {{if this.isLarge "Smaller" "Larger"}}</small>
</button>
```

Whether this is an improvement in clarity is mostly a matter of taste.

Either way, we have significantly reduced the duplication and made the
important bits stand out from the rest. Running the test suite confirms
everything is still working as expected. Great!