Add `<Map>` component
Let's generate a new component for our map.

```
ember generate component --with-component-class map
```

Since not every component has behavior associated with them, the
component generator does not generate a JavaScript file by default. As
we saw earlier, we can always use the `component-class` generator to add
one later.

However, in this case, we are pretty sure we will need one, we can also
pass the `--with-component-class` flag to generator so that we have
everything we need from the get-go.

> Zoey says: Too much typing? Use `ember g component -gc map` instead.
> That's a saving of 26 characters, or exactly one alphabet song's worth
> of characters! The `-gc` flag stands for **G**limmer **c**omponent, but
> you may also remember it as **g**enerate **c**lass.

Let's start with the JavaScript file:

```js {data-filename="app/components/map.js" data-diff="+2,+5,+6,+7"}
import Component from '@glimmer/component';
import ENV from 'super-rentals/config/environment';

export default class MapComponent extends Component {
  get token() {
    return encodeURIComponent(ENV.MAPBOX_ACCESS_TOKEN);
  }
}
```

We imported the access token from the config file and returned it from a
`token` _getter_. This allows us to access it as `this.token` on the
component instance, including from the template. It is also important to
URL-encode the token, in case it contains any special characters that
are not URL-safe.

On the template side:

```handlebars {data-filename="app/components/map.hbs}
<div class="map">
  <img
    alt="Map image at coordinates {{@lat}},{{@lng}}"
    ...attributes
    src="https://api.mapbox.com/styles/v1/mapbox/streets-v11/static/{{@lng}},{{@lat}},{{@zoom}}/{{@width}}x{{@height}}@2x?access_token={{this.token}}"
    width={{@width}} height={{@height}}
  >
</div>
```

First, we have a container element for styling purposes.

Then, we have an `<img>` tag to request and render the static map image
from mapbox.

We generated a reasonable default value for the `alt` attribute` based
on the value of `{{@lat}}` and `{{@lng}}`. These are _arguments_ that
the invoker would need to specify when invoking the `<Map>` component.

By _parameterizing_ our component using arguments, we made a reusable
components that can be invoked from different parts of the app and
customized to meet the needs for those specific contexts.

We have already seen this in action when using the `<LinkTo>` component
earlier; we had to specify a `@route=...` argument so that it knows what
page to navigate to.

You may also notice that we are directly _interpolating_ values into the
alt attribute's value. Ember will automatically concatenate them into
the final string value for us, including doing the appropiate escaping
if necessary.

Next, we used `...attributes` to allow the invoker to further
customization of the `<img>` tag, such as passing extra attributes such
as `class`, as well as _overriding_ our default `alt` attribute with a
more specific or human-friendly one.

*The ordering is important here!* Ember apply the attributes in the
order the appear in. By assigning the `alt` attribute first before
`...attributes` is applied, we are explicitly allowing the invoker to
optionally pass a different value for the `alt` attribute.

Since the passed-in `alt` attribute (if any) will appear _after_ ours,
it will override the value we specified. On the other hand, it is
important that we assign `src`, `width` and `height` after
`...attributes`, so that they don't get accidentally overwritten by the
invoker.

The `src` attribute interpolates together all the required parameters
into the URL format for mapbox's [static map image
API](https://docs.mapbox.com/api/maps/#static-images), including the
escaped access token from `this.token`.

Finally, since we are using the `@2x` "retina" image, we should specify
the `width` and `height` attributes. Otherwise, the `<img>` will be
rendered at twice the size than what we expected.

That's a lot of behaviors, let's write some tests! In particular, we
should make sure to have some _test coverage_ for the HTML attributes
overriding behavior we discussed above.

<!-- TODO: show diff -->

*Finger crossed...* Hey, all the tests passed! But does that mean it
actually works in practice? Let's find out!

<!-- TODO: show diff -->

Hey! That's a map!

> Zoey says: If the map image failed to load, make sure you have the
> correct `MAPBOX_ACCESS_TOKEN` in `config/environment.js`. Don't forget
> to restart the development and test servers after editing your config
> file!

<!-- TODO: track https://github.com/ember-cli/ember-cli/issues/8782 -->

Finally, we will add an assertion to the `<Rental>` tests to make sure
we rendered the `<Map>` component successfully.

<!-- TODO: show diff -->